% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{av_exe_and_tax_code_by_pin}
\alias{av_exe_and_tax_code_by_pin}
\title{Assessed values (AVs), exemptions, and tax codes by PIN and year}
\format{
A data frame containing 11233251 rows and 15 columns:
\describe{
  \item{year}{Tax year}
  \item{pin}{Property index number identifying a specific parcel or unit}
  \item{tax_code}{5-digit tax code}
  \item{class}{Property class}
  \item{av}{Assessed value}
  \item{exe_homeowner}{Homeowner exemption}
  \item{exe_senior}{Senior exemption}
  \item{exe_freeze}{Senior freeze exemption}
  \item{exe_longtime_homeowner}{Longtime homeowner exemption}
  \item{exe_disabled}{Persons with disabilities exemption}
  \item{exe_vet_returning}{Returning veterans exemption}
  \item{exe_vet_dis_lt50}{Veteran less than 50\% disability}
  \item{exe_vet_dis_50_69}{Veteran greater than 50\% and
  less than 69\% disability}
  \item{exe_vet_dis_ge70}{Veteran greater than 70\% disability}
  \item{exe_abate}{Abatement exemption}
}
}
\source{
CLERKVALUES AS/400 table (or equivalent SQL table)
}
\usage{
av_exe_and_tax_code_by_pin
}
\description{
PIN-level assessments and exemptions pulled from internal CCAO
  databases. These values are used in lookup functions and when calculating
  tax bills.

  Note that exemption columns store the *amount* of the exemption, rather
  than a binary did/did not receive indicator. Amounts are recorded in *EAV*,
  rather than the tax-value of the exemption.

  Also contains each PIN's tax code, which defines the unique combination of
  taxing districts for a given property/location. All properties in a tax
  code will share the same tax districts and rates.

  New tax codes are created when new taxing districts (including TIFS and
  SSAs) are established.
}
\keyword{datasets}
