% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_levies.R
\name{change_levies}
\alias{change_levies}
\alias{change_levies_amt}
\alias{change_levies_avg}
\alias{change_levies_cpi}
\alias{change_levies_pct}
\title{Change levy values to test counterfactual scenarios}
\usage{
change_levies_amt(levies_df, amount, ...)

change_levies_avg(levies_df, n_years, ...)

change_levies_cpi(levies_df, ...)

change_levies_pct(levies_df, percent, ...)
}
\arguments{
\item{levies_df}{A data frame of levy information identical to the format
returned by \code{\link{lookup_levies}}.}

\item{amount}{A numeric vector of amounts to be added to the levies in
\code{levies_df}. Must be length 1 or the same number of rows as
\code{levies_df}.}

\item{...}{Arguments passed directly to the underlying PTELL check function
\code{\link{check_levies_ptell}}.}

\item{n_years}{A length 1 numeric vector specifying the number of prior year
percentage changes to average. For example, providing levies from 2019 and
specifying \code{n_years = 5} will get the average levy percent change
between 2014 and 2019, then multiply the 2019 levy by that average.}

\item{percent}{A numeric vector of percentages by which to change the levies
in \code{levies_df}. Levies are changed via the following formula:
\code{levy + (percent * levy)}. Must be length 1 or the same number of rows
as \code{levies_df}.}
}
\value{
A data frame containing altered levy values in the same format
returned by \code{\link{lookup_levies}}.
}
\description{
The functions in this group alter levy data returned by
\code{\link{lookup_levies}} to simulate counterfactual situations.

For example, \code{\link{change_levies_pct}} can be combined with
\code{\link{tax_bill}} to determine how a tax bill would be affected if
all levies for that bill went up by 2\%.
}
\details{
Levy changes are subject to Illinois' Property Tax
Extension Limitation Law (PTELL). By default, these functions will throw a
warning when one or more levy changes might violate PTELL. Pass the
argument \code{quiet = TRUE} to disable this behavior.

Note that is is a naive/simple implementation of PTELL. See the underlying
\code{\link{check_levies_ptell}} function for more details on PTELL and
this specific implementation.
}
\section{Functions}{
\itemize{
\item \code{change_levies_amt}: Change levies by a fixed amount. The amount
specified is added to the existing levy.

\item \code{change_levies_avg}: Change levies by the average change from the
previous N years.

\item \code{change_levies_cpi}: Change levies by the prior year's rate of inflation
(CPI-U).

\item \code{change_levies_pct}: Change levies by a fixed percentage.
}}

\examples{
\donttest{
library(magrittr)
levies <- lookup_levies(2019, "73105")

levies \%>\%
  change_levies_amt(amount = 10000)
levies \%>\%
  change_levies_avg(n_years = 3)
levies \%>\%
  change_levies_cpi()
levies \%>\%
  change_levies_pct(percent = 0.02)
}
}
\seealso{
Other change: 
\code{\link{change_year}()}
}
\concept{change}
