% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_year.R
\name{change_year}
\alias{change_year}
\alias{change_year_copy_latest}
\title{Change the year of tax component data}
\usage{
change_year(df, from_year, to_year)

change_year_copy_latest(df)
}
\arguments{
\item{df}{A data frame in the format generated by
\code{\link{lookup_agency_eavs}}, \code{\link{lookup_exemptions}},
\code{\link{lookup_levies}}, or \code{\link{lookup_tifs}}. Must have a
\code{year} column.}

\item{from_year}{Year to copy data from. Must be a length 1 numeric vector.}

\item{to_year}{Year to change copied data to. Must be a length
1 numeric vector.}
}
\value{
Data frame of copied data with \code{year} updated to the value of
\code{to_year}.
}
\description{
Calculating future tax bill requires future (unavailable) data.
This function enables future calculations/simulations by copying past data
into the future. It can be combined with other functions to simulate
complex future scenarios.
}
\section{Functions}{
\itemize{
\item \code{change_year_copy_latest}: A helper function equivalent to
\code{change_year(df, max_year, max_year + 1)}. Copies the latest year and
adds 1 year.
}}

\examples{
library(magrittr)

lookup_levies(2019, "73105") \%>\%
  change_year(from_year = 2019, to_year = 2020)
}
\seealso{
Other change: 
\code{\link{change_levies}}
}
\concept{change}
