% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_db_conn}
\alias{check_db_conn}
\title{Check if a connection object is a valid PTAXSIM database}
\usage{
check_db_conn(conn)
}
\arguments{
\item{conn}{A valid DBI connection object pointing to a local PTAXSIM
SQLite database file.}
}
\value{
Returns \code{TRUE} if all requirements are met, otherwise returns
an error.
}
\description{
Checks the size, table names, and validity of a DBI connection
object for the PTAXSIM database. Throws an error if the size or table names
are not as expected.
}
\examples{
\dontrun{
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db")
check_db_conn(ptaxsim_db_conn)
}
}
\seealso{
Other check: 
\code{\link{check_db_sync}()}
}
\concept{check}
