% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_db_sync}
\alias{check_db_sync}
\title{Check if the PTAXSIM package and database file are synchronized}
\usage{
check_db_sync(conn)
}
\arguments{
\item{conn}{A valid DBI connection object pointing to a local PTAXSIM
database file.}
}
\value{
Returns \code{TRUE} if all requirements are met, otherwise returns
an error.
}
\description{
Performs a two-way check to synchronize the package and
database.
\enumerate{
\item Package version must be \code{>=} the minimum package version recorded
in the database \code{metadata} table
\item Database version must be \code{>=} the minimum database version recorded
in the package \code{DESCRIPTION} file
}
}
\examples{
\dontrun{
ptaxsim_db_conn <- DBI::dbConnect(RSQLite::SQLite(), "./ptaxsim.db")
check_db_sync(ptaxsim_db_conn)
}
}
\seealso{
Other check: 
\code{\link{check_db_conn}()}
}
\concept{check}
