% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_levies_ptell}
\alias{check_levies_ptell}
\title{Check that levy changes conform to PTELL}
\usage{
check_levies_ptell(
  levies_df_old,
  levies_df_new,
  quiet = FALSE,
  agency_eavs_df = ptaxsim::levy_and_total_eav_by_agency
)
}
\arguments{
\item{levies_df_old}{A data frame in the format returned by
\code{\link{lookup_levies}}. Contains the original (unchanged) levies.}

\item{levies_df_new}{A data frame in the format returned by
\code{\link{lookup_levies}}. Contains the new, altered levies.}

\item{quiet}{Default FALSE. A boolean value for whether or not to warn about
violations of PTELL. See details for more information.}

\item{agency_eavs_df}{A data frame containing all aggregate agency EAVs to be
used to calculate PTELL limiting rates. Defaults to the actual EAVs
supplied by the Clerk in \code{\link{levy_and_total_eav_by_agency}}. Can be
altered in the case of changed or future (estimated EAVs).}
}
\value{
Returns a warning when any change to levies between
\code{levies_df_old} and \code{levies_df_new} could violate PTELL.
}
\description{
The Property Tax Extension Limitation Law (PTELL) limits the
growth of taxing agency levies. It caps the maximum levy increase for a
given year at the rate of inflation or 5\%, whichever is less. PTELL is
colloquially known as "tax caps."

All non-home rule taxing districts in Cook County are subject to PTELL.
This typically includes smaller government agencies such as libraries,
forest preserves, and townships.

PTELL uses the prior year's urban consumer price index (CPI-U) to calculate
the inflation rate. For example, the tax year 2020 CPI-U was 2.3\%, based on
the change from December 2018 to December 2019.

The goal of this function is to check that hypothetical levy changes do not
violate PTELL. However, the true application of PTELL is extremely complex.
As such, this function aims to be a heuristic, \emph{not a true check}. It comes
with the following caveats, use at your own discretion:
\itemize{
\item Ignores home rule districts via home rule indicators provided by the
Cook County Clerk's Office
\item Does not perform rate limits checks for specific funds (this
functionality is entirely excluded from PTAXSIM)
\item Ignores potential Clerk adjustments to final levies to account for loss
\item Does not differentiate between cap and non-cap funds
\item Certain property types are supposed to be excluded when calculating PTELL
limiting rates, including:
\itemize{
\item New property
\item Property annexed from other districts
\item Recovered TIF increment (after expiration)
\item Expired incentives that reduce AV or EAV
\item Disconnected property
}

This function ignores those exclusions and simply uses the aggregate EAV
for each taxing district provided by the Clerk
\item In practice, the Clerk uses the highest aggregate levy of the past
3 years to calculate the limiting rate. However, this function simply
uses the prior year's levy
}
}
\concept{check}
