% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equalization_factors}
\alias{equalization_factors}
\title{State equalization factor by year}
\format{
A data frame containing 41 rows and 2 columns:
\describe{
  \item{year}{Tax year}
  \item{equalization_factor}{State multiplier/equalization factor}
}
}
\source{
\url{https://www.cookcountyclerkil.gov/publications/equalization-factor-overview-1973-current-0}
}
\usage{
equalization_factors
}
\description{
The state multiplier/equalization factor is calculated by the
  Illinois Department of Revenue. It is a multiplier specifically for Cook
  County assessments. Cook County assesses property between 10 and 25% of
  fair market value, while most other counties assess at the state-mandated
  33.33%.

  The equalizer multiplies Cook County aggregate assessed values to bring
  them up to the state-mandated 33.33% level of assessment.
}
\keyword{datasets}
