% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_df}
\alias{lookup_df}
\alias{lookup_agency_eavs}
\alias{lookup_exemptions}
\alias{lookup_levies}
\alias{lookup_tifs}
\title{Lookup data frames required to calculate property tax bills}
\usage{
lookup_agency_eavs(year, tax_code)

lookup_exemptions(year, pin)

lookup_levies(year, tax_code)

lookup_tifs(year, tax_code)
}
\arguments{
\item{year}{A numeric vector of tax years.}

\item{tax_code}{A character vector of 5-digit Cook County tax codes.}

\item{pin}{A character vector of 14-digit PINs.}
}
\value{
A data frame containing the looked-up values.
}
\description{
The functions in this group take a tax year, PIN, and/or tax
code as input and return a data frame as an output. The returned data
frames are in a specific format (in terms of column name/order) and can be
used directly in \code{\link{tax_bill}}.

When calculating tax bills for future years or providing custom/simulated
data, you'll need to replicate the format returned by these functions.

Note: All vector inputs must be either the same length or length 1.
}
\section{Functions}{
\itemize{
\item \code{lookup_agency_eavs}: Lookup tax agency (CPS, Cook County, forest preserve)
total equalized assessed values. Such values are known as the "base" and
are used to calculate per agency tax rates. Returns a data frame with 1 row
for each agency associated with the input tax code and year.

\item \code{lookup_exemptions}: Lookup exemptions granted to a PIN for the specified
year. Returns a data frame with 1 row per PIN per year.

\item \code{lookup_levies}: Lookup tax agency (CPS, Cook County, forest preserve)
total levies. Returns a data frame with 1 row for each agency associated
with the input tax code and year.

\item \code{lookup_tifs}: Lookup any TIFs that apply to a given tax code and
year. Returns a data frame with only 1 row per tax code and year, or 0 rows
if the tax code is not within a TIF.
}}

\examples{
\donttest{
lookup_agency_eavs(2019, "73105")
lookup_exemptions(2018:2019, "20304190020000")
lookup_levies(2019, "73105")
lookup_tifs(2019, "73105")
}
}
\seealso{
Other lookups: 
\code{\link{lookup_vec}}
}
\concept{lookups}
