% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_dt}
\alias{lookup_dt}
\alias{lookup_agency}
\alias{lookup_pin}
\alias{lookup_tif}
\title{Lookup data sets required to calculate property tax bills}
\usage{
lookup_agency(year, tax_code, conn = ptaxsim_db_conn)

lookup_pin(year, pin, stage = "clerk", conn = ptaxsim_db_conn)

lookup_tif(year, tax_code, conn = ptaxsim_db_conn)
}
\arguments{
\item{year}{A numeric vector of tax years.}

\item{tax_code}{A character vector of 5-digit Cook County tax codes.}

\item{conn}{A connection object pointing to a local copy of the
PTAXSIM database.}

\item{pin}{A character vector of 14-digit PINs.}

\item{stage}{A length 1 character vector indicating the assessment stage
from which to pull assessed value (column \code{av}) and equalized assessed
value (column \code{eav}). Options include \code{"mailed"},
\code{"certified"}, \code{"board"}, and \code{"clerk"}.}
}
\value{
A \code{data.table} containing the specified tax data.
}
\description{
The functions in this group take a tax year, PIN, and/or tax
code as input and return a \code{data.table} as an output. The returned
data are in a specific format (in terms of column name/type/order)
and can be used directly in \code{\link{tax_bill}} or piped to helper
functions.

When calculating tax bills for future years or providing custom/simulated
data, you'll need to replicate the format returned by these functions.
}
\section{Functions}{
\itemize{
\item \code{lookup_agency()}: Lookup tax agency (CPS, Cook County, forest preserve)
total equalized assessed values. Such values are known as the "base" and
are used to calculate per agency tax rates. Also returns each agency's
levy amount. Returns a \code{data.table} with 1 row for each agency
associated with the input tax code and year.

\item \code{lookup_pin()}: Lookup the equalized assessed value and exemptions
for a specific PIN and year. Returns a \code{data.table} with AV, EAV, and
a column specifying the EAV value of each received exemption.

\item \code{lookup_tif()}: Lookup any TIFs that apply to a given tax code and
year. Returns a \code{data.table} with only 1 row per tax code and year,
or 0 rows if the tax code is not part of a TIF.

}}
\examples{
\dontrun{
lookup_agency(2019, "73105")
lookup_pin(2018:2019, "20304190020000")
lookup_tif(2019, "73105")
}
}
\seealso{
Other lookups: 
\code{\link{lookup_vec}}
}
\concept{lookups}
