% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_geom}
\alias{lookup_geom}
\alias{lookup_pin10_geometry}
\title{Lookup boundaries associated with taxing entities}
\usage{
lookup_pin10_geometry(year, pin10, conn = ptaxsim_db_conn)
}
\arguments{
\item{year}{A numeric vector of tax years.}

\item{pin10}{A character vector of 10-digit PINs.}

\item{conn}{A connection object pointing to a local copy of the
PTAXSIM database.}
}
\value{
A \code{data.table} containing the specified geometry. NOTE: The
geometry is returned as a character vector in the Well-Known Text (WKT)
format. You must convert the character vector before using \code{sf} or
similar packages to perform spatial manipulation.
}
\description{
The functions in this group take a tax year and identifier as
input and return a \code{data.table} as an output. The returned data
contains a geometry column representing the spatial boundary of the taxing
entity.
}
\section{Functions}{
\itemize{
\item \code{lookup_pin10_geometry()}: Lookup the parcel boundary for a specific PIN and
year. Returns a \code{data.table} with the PIN's geometry (as WKT) as well
as the longitude and latitude of the PIN's centroid. NOTE: Not all PINs
in the tax bill database have corresponding geometries.

}}
\examples{
\dontrun{
lookup_pin10_geometry(2019, "2030419002")
}
}
\seealso{
Other lookups: 
\code{\link{lookup_dt}},
\code{\link{lookup_vec}}
}
\concept{lookups}
