% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_vec}
\alias{lookup_vec}
\alias{lookup_eav}
\alias{lookup_equalization_factor}
\alias{lookup_tax_code}
\title{Lookup values required to calculate property tax bills}
\usage{
lookup_eav(year, pin)

lookup_equalization_factor(year)

lookup_tax_code(year, pin)
}
\arguments{
\item{year}{A numeric vector of tax years.}

\item{pin}{A character vector of 14-digit PINs.}
}
\value{
A vector containing the looked-up values. Will be the same length as
the longest vector in the inputs.
}
\description{
The functions in this group take a tax year and PIN as input and
return a vector as an output. The output vector can be used directly in
\code{\link{tax_bill}}.

Note: All vector inputs must be either the same length or length 1.
}
\section{Functions}{
\itemize{
\item \code{lookup_eav}: Lookup the equalized assessed value for a specific PIN
and year. Returns a numeric vector the same length as the longest input.

\item \code{lookup_equalization_factor}: Lookup the state equalization rate for a given year.
Returns a numeric vector the same length as the input.

\item \code{lookup_tax_code}: Lookup 5-digit Cook County tax code by PIN and year.
Returns a character vector the same length as the longest input.
}}

\examples{

lookup_eav(2017:2019, "20304190020000")
lookup_equalization_factor(2014:2019)
lookup_tax_code(2019, c("20304190020000", "16152090350000"))
}
\seealso{
Other lookups: 
\code{\link{lookup_df}}
}
\concept{lookups}
