% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tax_bills_detail}
\alias{sample_tax_bills_detail}
\title{Detailed data from real sample tax bills}
\format{
A data frame containing 381 rows and 11 columns:
\describe{
  \item{year}{Tax year}
  \item{pin}{Property index number identifying a specific parcel or unit}
  \item{class}{Property class}
  \item{agency}{Tax district/agency ID number}
  \item{agency_name}{Tax district/agency name. NOTE: This may be different
  from the name produced by \code{\link{tax_bill}}, since it is copied
  directly from the PDF tax bill}
  \item{tax}{Total tax amount after exemptions for the specified agency}
  \item{rate}{Total tax rate for the specified agency in this tax code}
  \item{percent}{Percent of the total tax bill for this line item}
  \item{pension}{Dollar amount from this line item dedicated to pensions}
  \item{prev_tax}{Total tax amount after exemptions for the specified agency
  from the previous year}
  \item{tif_total}{Total amount sent to TIFs for this tax bill}
}
}
\source{
\url{https://www.cookcountytreasurer.com/setsearchparameters.aspx}
}
\usage{
sample_tax_bills_detail
}
\description{
Data frame of the summary table from each tax bill in
  \code{data-raw/}. Each bill is scanned with Tabula and then formatted to
  match the output of \code{\link{tax_bill}}. The line item values are used
  for unit/integration testing.
}
\keyword{datasets}
