% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tax_bills_summary}
\alias{sample_tax_bills_summary}
\title{Summary data from real sample tax bills}
\format{
A data frame containing 31 rows and 17 columns:
\describe{
  \item{year}{Tax year}
  \item{pin}{Property index number identifying a specific parcel or unit}
  \item{tax_code}{5-digit tax code}
  \item{class}{Property class}
  \item{township_name}{Township name}
  \item{in_tif}{Boolean for whether the PIN/tax code is in a Tax Increment
  Finance (TIF) district}
  \item{in_ssa}{Boolean for whether the PIN/tax code is in a Special Service
  Area (SSA)}
  \item{exe_homeowner}{Homeowner exemption}
  \item{exe_senior}{Senior exemption}
  \item{exe_freeze}{Senior freeze exemption}
  \item{exe_longtime_homeowner}{Longtime homeowner exemption}
  \item{exe_disabled}{Persons with disabilities exemption}
  \item{exe_vet_returning}{Veterans exemption}
  \item{exe_vet_disabled}{Veterans disability exemption}
  \item{eav}{Total EAV before exemptions}
  \item{tax_amt_pre_exe}{Total tax amount before exemptions}
  \item{tax_amt_post_exe}{Total tax amount after exemptions}
}
}
\source{
\url{https://www.cookcountytreasurer.com/setsearchparameters.aspx}
}
\usage{
sample_tax_bills_summary
}
\description{
The Cook County Treasurer is responsible for producing the
  actual tax bills sent to taxpayers. This package contains a set of real
  sample tax bills (stored in \code{/data-raw} as PDFs). This data set
  contains the summary statistics and metadata from those tax bills.

  Note: All exemption columns (prefixed with \code{exe_}) are booleans
  for whether or not that exemption was received.
}
\keyword{datasets}
