% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_bill.R
\name{tax_bill}
\alias{tax_bill}
\title{Calculate an approximate Cook County property tax bill}
\usage{
tax_bill(
  year_vec,
  pin_vec,
  tax_code_vec = lookup_tax_code(year_vec, pin_vec),
  pin_df = lookup_pin(year_vec, pin_vec),
  agency_df = lookup_agency(year_vec, tax_code_vec),
  tif_df = lookup_tif(year_vec, tax_code_vec),
  simplify = TRUE
)
}
\arguments{
\item{year_vec}{Numeric vector of tax years for which to return bills.}

\item{pin_vec}{Character vector of 14-digit Property Index Numbers (PINs)
with no dashes or spaces.}

\item{tax_code_vec}{Character vector of 5-digit Cook County tax codes. These
codes are included on individual property tax bills. If missing, the
\code{\link{lookup_tax_code}} function will be used to retrieve tax codes
based on \code{year_vec} and \code{pin_vec}.}

\item{pin_df}{Data frame containing the exemptions and assessed value
specific to each PIN and year. Data \emph{must} be identical to the format
returned by \code{\link{lookup_pin}}. If missing, \code{\link{lookup_pin}}
will be used to retrieve each PIN's information based on \code{year_vec}
and \code{pin_vec}.}

\item{agency_df}{Data frame containing the levy and base amount for each
taxing district in the specified tax code. Data \emph{must} be identical to the
format returned by \code{\link{lookup_agency}}. If missing,
\code{\link{lookup_agency}} will be used to retrieve each taxing
jurisdiction's levies based on \code{year_vec} and \code{tax_code_vec}.}

\item{tif_df}{Data frame containing any TIF applicable to the specified
tax code. Will be an empty data frame if no TIF exists for the property.
Data \emph{must} be identical to the format returned by
\code{\link{lookup_tif}}. If missing, \code{\link{lookup_tif}} will be
used to retrieve the tax code's TIF share based on \code{year_vec}
and \code{tax_code_vec}.}

\item{simplify}{Default TRUE. Boolean for whether or not to keep only the
the most vital columns in the final output, including tax amounts, rates,
and amounts to TIFs.}
}
\value{
A data frame which contains a tax bill for each specified PIN and
year. Each tax bill is broken out by taxing district, meaning there is a
row for each district relevant to each PIN. Most PINs have 10-15 districts
(rows) associated with each year.

Note that unlike official tax bills, TIFs are not broken out into a
separate line item and instead are contained in their own column
(named \code{final_tax_to_tif}).
}
\description{
This function calculates a Cook County property tax bill using
levy, base, and TIF information from the Cook County Clerk and Assessor.
It allows users to simulate different tax scenarios by changing the input
data. For example, this function can answer the following questions:
\itemize{
\item What would my tax bill be if my assessed value was $10K lower?
\item What would my tax bill be if Chicago increased its levy by 5\%?
\item What would my tax bill be if my property was in a different neighborhood?
\item What would my tax bill be if a large new development was added to
my neighborhood?
}

Most of the data necessary to answer these questions is built into the
package. Data for the most current tax year is usually available the
following year (2020 data is available in 2021). If needed, users can
supply current tax year data manually. See vignettes for more information.
}
\details{
Note that all vector inputs (suffixed with \code{_vec}) have two
input modes:
\enumerate{
\item If the input vectors are the same length, pairwise tax bills are
returned (each PIN, year, and tax code combination returns 1 bill).
\item If the input vectors are different lengths, the Cartesian product of
the vectors is returned (5 years and 10 PINs will return all 5 years' of
bills for each PIN).
}

The district-level tax amounts returned by this function will \emph{not}
perfectly match the amounts on real tax bills. This is due to rounding
and truncating that occurs in the real system. Most estimated amounts will
still be within a few dollars of the amounts on real bills.
}
\examples{
\dontrun{
# Get a single tax bill
tax_bill(2019, "14081020210000")

# Get tax bills for multiple years
tax_bill(2017:2019, "14081020210000")

# Get multiple tax bills for different PINs and years
tax_bill(c(2014, 2019), c("14081020210000", "07133020190000"))

# Get a tax bill for a new location
tax_bill(2019, "14081020210000", tax_code_vec = "35011")
}
}
\concept{tax_bill}
