% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tax_codes_by_agency}
\alias{tax_codes_by_agency}
\title{Cook County taxing agency/district by tax code and year}
\format{
A data frame containing 303980 rows and 6 columns:
\describe{
  \item{year}{Tax year}
  \item{tax_code}{5-digit tax code}
  \item{agency}{Tax district/agency ID number}
}
}
\source{
\url{https://www.cookcountyclerkil.gov/service/tax-extension-and-rates}
}
\usage{
tax_codes_by_agency
}
\description{
Each Cook County taxing district (usually) spans multiple tax
  codes. Similarly, each tax code usually contains multiple overlapping
  taxing districts. This table is a many-to-many lookup of tax codes to
  taxing districts.

  Tax codes and taxing districts can change year-to-year as new agencies are
  created or deprecated.
}
\keyword{datasets}
