% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tifs}
\alias{tifs}
\title{Tax Increment Financing (TIF) district information by tax code and year}
\format{
A data frame containing 7792 rows and 19 columns:
\describe{
  \item{year}{Tax year}
  \item{tax_code}{5-digit tax code}
  \item{agency}{Tax district/agency ID number of the TIF. TIFs are recorded
  as separate agencies within the Clerk's system}
  \item{tif_name}{TIF district common name}
  \item{tax_code_rate}{Aggregate tax code tax rate, sum of all agency rates
  for agencies in the tax code}
  \item{tax_code_eav}{Total tax code EAV for the year}
  \item{tax_code_frozen_eav}{Total tax code EAV for the year the TIF was
  created}
  \item{tax_code_increment}{Current tax code EAV above the frozen amount}
  \item{tax_code_revenue}{Tax revenue generated for the TIF from the
  increment in this tax code for the year}
  \item{tax_code_distribution_percent}{The percent of total EAV in the
  increment/above the frozen amount. Also known as the TIF share}
  \item{tif_total_eav}{Total EAV for the entire TIF for the year}
  \item{tif_total_frozen_eav}{Total EAV for the entire TIF for the year
  the TIF was created}
  \item{tif_total_increment}{Current total TIF EAV above the frozen amount}
  \item{tif_total_revenue}{Tax revenue generated for the entire TIF for
  the year}
  \item{first_year}{Year the TIF was created}
  \item{this_year_revenue}{Tax revenue generated for the entire TIF for
  the year}
  \item{prev_year_revenue}{Tax revenue generated for the entire TIF for
  the previous year}
  \item{tif_new_this_year}{Boolean for first year of the TIF}
  \item{tif_cancelled_this_year}{Boolean for final year of the TIF. Some TIFS
  are cancelled early}
}
}
\source{
\url{https://www.cookcountyclerkil.gov/service/tif-reports}
}
\usage{
tifs
}
\description{
Tax Increment Financing districts (commonly referred to as TIFS)
  are a taxing tool used to generate revenue for the economic development of
  a specific area.

  They work by capping the EAV of an area, then capturing any taxes on value
  above the "frozen" EAV (the increment).
}
\keyword{datasets}
