% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{levy_and_total_eav_by_agency}
\alias{levy_and_total_eav_by_agency}
\title{Levy and aggregate (base) equalized assessed value by tax agency by year}
\format{
A data frame containing 6874 rows and 6 columns:
\describe{
  \item{year}{Tax year}
  \item{agency}{Tax district/agency ID number}
  \item{agency_name}{Tax district/agency name}
  \item{agency_rate}{Agency tax rate for this year. Equal to levy/base}
  \item{total_eav}{Total equalized assessed value (EAV) in the taxing
  district. Also known as the base}
  \item{total_levy}{Total levy amount the taxing district requested}
}
}
\source{
\url{https://www.cookcountyclerkil.gov/service/tax-extension-and-rates}
}
\usage{
levy_and_total_eav_by_agency
}
\description{
Cook County is made up of numerous independent taxing bodies
  (AKA agencies or districts) that all receive funding from property taxes.
  The amount each district receives is determined by its levy. The pool of
  available value each district can tax is known as its base/total EAV.

  Each district's individual tax rate is determined by dividing its levy by
  its base EAV.
}
\keyword{datasets}
