% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_vec}
\alias{lookup_vec}
\alias{lookup_tax_code}
\title{Lookup values required to calculate property tax bills}
\usage{
lookup_tax_code(year, pin, conn = ptaxsim_db_conn)
}
\arguments{
\item{year}{A numeric vector of tax years.}

\item{pin}{A character vector of 14-digit PINs.}

\item{conn}{A connection object pointing to a local copy of the
PTAXSIM database.}
}
\value{
A vector containing the specified tax data values.
}
\description{
The functions in this group take a tax year and PIN as input and
return a vector as an output. The output vector can be used directly in
\code{\link{tax_bill}}.
}
\section{Functions}{
\itemize{
\item \code{lookup_tax_code()}: Lookup the Cook County tax code for a specific PIN and
year. Tax codes represent the unique geographic overlap of different
taxing districts. If the input vectors are both length N, then the output
will also be length N. If the input vectors are different lengths, the
Cartesian product is returned.

}}
\examples{
\dontrun{
lookup_tax_code(2019, c("20304190020000", "16152090350000"))
lookup_tax_code(2006:2021, c("20304190020000", "16152090350000"))
}
}
\seealso{
Other lookups: 
\code{\link{lookup_dt}},
\code{\link{lookup_geom}}
}
\concept{lookups}
